Fundamental Kotlin - Exceptions.
